style: 통역 시작 화면 ui 구현 [FE-63]
