+++ src/app/routes/Routes.tsx
@@ -12,7 +12,6 @@ import {
   SignupSuccessPage,
   NotFoundPage,
   MapPage,
-  TranslateGuidePage,
 } from "@/pages";
 
 import MedicalRecordList from "@/pages/records/medicalrecord-list";
@@ -71,7 +70,6 @@ export default function AppRoutes() {
 
         {/* 통역 페이지 */}
         <Route path="/translate" element={<TranslatePage />} />
-        <Route path="/translate/guide" element={<TranslateGuidePage />} />
 
         {/* 마이 페이지 */}
         <Route path="/mypage" element={<Mypage />} />
